# openapi/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "wearable-analytics-api",
    "version": "0.1.0",
    "description": "OpenAPI specification for the wearable analytics API service.",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local development server"
    }
  ],
  "security": [],
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "OverviewResponse": {
        "type": "object",
        "properties": {
          "generatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "day": {
            "type": "string",
            "format": "date-time"
          },
          "metrics": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "label": {
                  "type": "string"
                },
                "value": {
                  "type": "string"
                },
                "detail": {
                  "type": "string"
                }
              },
              "required": [
                "label",
                "value",
                "detail"
              ]
            }
          },
          "notes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "anomalies": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "severity": {
                  "type": "string",
                  "enum": [
                    "low",
                    "medium",
                    "high"
                  ]
                },
                "title": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                }
              },
              "required": [
                "severity",
                "title",
                "description"
              ]
            }
          },
          "sync": {
            "type": "object",
            "properties": {
              "running": {
                "type": "boolean"
              },
              "latestStatus": {
                "anyOf": [
                  {
                    "type": "string",
                    "enum": [
                      "pending",
                      "running",
                      "succeeded",
                      "failed"
                    ]
                  },
                  {
                    "type": "string",
                    "enum": [
                      "idle"
                    ]
                  }
                ]
              },
              "lastSyncedAt": {
                "type": "string",
                "nullable": true,
                "format": "date-time"
              }
            },
            "required": [
              "running",
              "latestStatus",
              "lastSyncedAt"
            ]
          },
          "connection": {
            "type": "object",
            "properties": {
              "connected": {
                "type": "boolean"
              },
              "configured": {
                "type": "boolean"
              },
              "needsReconnect": {
                "type": "boolean"
              }
            },
            "required": [
              "connected",
              "configured",
              "needsReconnect"
            ]
          }
        },
        "required": [
          "generatedAt",
          "day",
          "metrics",
          "notes",
          "anomalies",
          "sync",
          "connection"
        ]
      },
      "BaselineSnapshotResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "day": {
            "type": "string",
            "format": "date-time"
          },
          "hrvBaseline": {
            "type": "number",
            "nullable": true
          },
          "restingHrBaseline": {
            "type": "number",
            "nullable": true
          },
          "temperatureBaseline": {
            "type": "number",
            "nullable": true
          },
          "sleepDurationBaseline": {
            "type": "number",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "userId",
          "day",
          "hrvBaseline",
          "restingHrBaseline",
          "temperatureBaseline",
          "sleepDurationBaseline",
          "createdAt",
          "updatedAt"
        ]
      },
      "RecoveryScoreResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "day": {
            "type": "string",
            "format": "date-time"
          },
          "score": {
            "type": "integer"
          },
          "confidence": {
            "type": "number",
            "nullable": true
          },
          "hrvContribution": {
            "type": "number",
            "nullable": true
          },
          "restingHrContribution": {
            "type": "number",
            "nullable": true
          },
          "temperatureContribution": {
            "type": "number",
            "nullable": true
          },
          "sleepContribution": {
            "type": "number",
            "nullable": true
          },
          "explanationSummary": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "userId",
          "day",
          "score",
          "confidence",
          "hrvContribution",
          "restingHrContribution",
          "temperatureContribution",
          "sleepContribution",
          "explanationSummary",
          "createdAt",
          "updatedAt"
        ]
      },
      "RecoveryDetailResponse": {
        "type": "object",
        "properties": {
          "day": {
            "type": "string",
            "format": "date-time"
          },
          "score": {
            "type": "integer"
          },
          "confidence": {
            "type": "number",
            "nullable": true
          },
          "explanationSummary": {
            "type": "string"
          },
          "vitals": {
            "type": "object",
            "properties": {
              "hrv": {
                "type": "number",
                "nullable": true
              },
              "restingHeartRate": {
                "type": "number",
                "nullable": true
              },
              "temperatureDeviation": {
                "type": "number",
                "nullable": true
              },
              "sleepSeconds": {
                "type": "integer"
              }
            },
            "required": [
              "hrv",
              "restingHeartRate",
              "temperatureDeviation",
              "sleepSeconds"
            ]
          },
          "factors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "key": {
                  "type": "string",
                  "enum": [
                    "hrv",
                    "restingHeartRate",
                    "sleepDuration",
                    "temperature"
                  ]
                },
                "label": {
                  "type": "string"
                },
                "contribution": {
                  "type": "number",
                  "nullable": true
                },
                "status": {
                  "type": "string",
                  "enum": [
                    "positive",
                    "negative",
                    "neutral"
                  ]
                },
                "currentValue": {
                  "type": "number",
                  "nullable": true
                },
                "baselineValue": {
                  "type": "number",
                  "nullable": true
                },
                "unit": {
                  "type": "string"
                },
                "detail": {
                  "type": "string"
                }
              },
              "required": [
                "key",
                "label",
                "contribution",
                "status",
                "currentValue",
                "baselineValue",
                "unit",
                "detail"
              ]
            }
          },
          "anomalies": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "severity": {
                  "type": "string",
                  "enum": [
                    "low",
                    "medium",
                    "high"
                  ]
                },
                "title": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                }
              },
              "required": [
                "severity",
                "title",
                "description"
              ]
            }
          }
        },
        "required": [
          "day",
          "score",
          "confidence",
          "explanationSummary",
          "vitals",
          "factors",
          "anomalies"
        ]
      },
      "SleepSummaryResponse": {
        "type": "object",
        "properties": {
          "day": {
            "type": "string",
            "format": "date-time"
          },
          "totals": {
            "type": "object",
            "properties": {
              "sleepSeconds": {
                "type": "integer"
              },
              "timeInBedSeconds": {
                "type": "integer"
              },
              "efficiency": {
                "type": "number",
                "nullable": true
              },
              "latencySeconds": {
                "type": "integer",
                "nullable": true
              }
            },
            "required": [
              "sleepSeconds",
              "timeInBedSeconds",
              "efficiency",
              "latencySeconds"
            ]
          },
          "vitals": {
            "type": "object",
            "properties": {
              "averageHr": {
                "type": "number",
                "nullable": true
              },
              "lowestHr": {
                "type": "number",
                "nullable": true
              },
              "averageHrv": {
                "type": "number",
                "nullable": true
              }
            },
            "required": [
              "averageHr",
              "lowestHr",
              "averageHrv"
            ]
          },
          "timing": {
            "type": "object",
            "properties": {
              "bedtimeStart": {
                "type": "string",
                "nullable": true,
                "format": "date-time"
              },
              "bedtimeEnd": {
                "type": "string",
                "nullable": true,
                "format": "date-time"
              }
            },
            "required": [
              "bedtimeStart",
              "bedtimeEnd"
            ]
          },
          "baseline": {
            "type": "object",
            "properties": {
              "sleepDurationSeconds": {
                "type": "number",
                "nullable": true
              },
              "durationDeltaSeconds": {
                "type": "number",
                "nullable": true
              }
            },
            "required": [
              "sleepDurationSeconds",
              "durationDeltaSeconds"
            ]
          },
          "consistency": {
            "type": "object",
            "nullable": true,
            "properties": {
              "averageClockMinutes": {
                "type": "integer"
              },
              "averageDeviationMinutes": {
                "type": "integer"
              },
              "latestOffsetMinutes": {
                "type": "integer"
              },
              "recent": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "day": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "bedtimeStart": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "clockMinutes": {
                      "type": "integer"
                    },
                    "offsetMinutes": {
                      "type": "integer"
                    }
                  },
                  "required": [
                    "day",
                    "bedtimeStart",
                    "clockMinutes",
                    "offsetMinutes"
                  ]
                }
              },
              "status": {
                "type": "string",
                "enum": [
                  "mixed",
                  "steady",
                  "variable",
                  "warming_up"
                ]
              }
            },
            "required": [
              "averageClockMinutes",
              "averageDeviationMinutes",
              "latestOffsetMinutes",
              "recent",
              "status"
            ]
          },
          "note": {
            "type": "string"
          }
        },
        "required": [
          "day",
          "totals",
          "vitals",
          "timing",
          "baseline",
          "consistency",
          "note"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/api/overview/latest": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "operationId": "getLatestOverview",
        "summary": "Get the latest overview summary",
        "responses": {
          "200": {
            "description": "Latest overview payload.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OverviewResponse"
                }
              }
            }
          },
          "404": {
            "description": "No overview data is available yet.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/baselines/latest": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "operationId": "getLatestBaseline",
        "summary": "Get the latest baseline snapshot",
        "responses": {
          "200": {
            "description": "Latest baseline snapshot.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaselineSnapshotResponse"
                }
              }
            }
          },
          "404": {
            "description": "No baseline data is available yet.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/recovery/latest": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "operationId": "getLatestRecoveryScore",
        "summary": "Get the latest recovery score",
        "responses": {
          "200": {
            "description": "Latest recovery score.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecoveryScoreResponse"
                }
              }
            }
          },
          "404": {
            "description": "No recovery data is available yet.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/recovery/latest/detail": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "operationId": "getLatestRecoveryDetail",
        "summary": "Get the latest detailed recovery summary",
        "responses": {
          "200": {
            "description": "Latest detailed recovery payload.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecoveryDetailResponse"
                }
              }
            }
          },
          "404": {
            "description": "No recovery detail is available yet.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/sleep/latest": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "operationId": "getLatestSleepSummary",
        "summary": "Get the latest sleep summary",
        "responses": {
          "200": {
            "description": "Latest sleep summary.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SleepSummaryResponse"
                }
              }
            }
          },
          "404": {
            "description": "No sleep summary is available yet.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/trends": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "operationId": "getTrendSummary",
        "summary": "Get the latest trend summary",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "7d",
                "30d"
              ]
            },
            "required": true,
            "name": "window",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Trend summary for the requested window.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrendSummaryResponse"
                }
              }
            }
          },
          "400": {
            "description": "The window query parameter is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No trend data is available yet.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/anomalies/recent": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "operationId": "getRecentAnomalies",
        "summary": "Get recent anomaly flags",
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 60
            },
            "required": false,
            "name": "limit",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Recent anomaly flags.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnomaliesCollectionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/anomalies/history": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "operationId": "getAnomalyHistory",
        "summary": "Get anomaly history grouped by day",
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 366
            },
            "required": false,
            "name": "limitDays",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "format": "date"
            },
            "required": false,
            "name": "cursorDay",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "format": "date"
            },
            "required": false,
            "name": "targetDay",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated anomaly history.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnomalyHistoryResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/anomalies/heatmap": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "operationId": "getAnomalyHeatmap",
        "summary": "Get anomaly heatmap data",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "3m",
                "6m",
                "12m"
              ]
            },
            "required": true,
            "name": "range",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "all",
                "sleep",
                "hrv",
                "resting_hr",
                "temperature"
              ]
            },
            "required": true,
            "name": "type",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Anomaly heatmap data.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnomalyHeatmapResponse"
                }
              }
            }
          },
          "400": {
            "description": "The range or type query parameter is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/ai/daily-brief": {
      "get": {
        "tags": [
          "AI"
        ],
        "operationId": "getAiDailyBrief",
        "summary": "Get the latest AI daily brief",
        "responses": {
          "200": {
            "description": "AI daily brief payload.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiDailyBriefResponse"
                }
              }
            }
          },
          "404": {
            "description": "No AI daily brief is available yet.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/ai/last-night": {
      "get": {
        "tags": [
          "AI"
        ],
        "operationId": "getAiLastNightSummary",
        "summary": "Get the latest AI sleep summary",
        "responses": {
          "200": {
            "description": "AI-oriented last-night summary.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiLastNightResponse"
                }
              }
            }
          },
          "404": {
            "description": "No AI last-night summary is available yet.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/ai/recovery": {
      "get": {
        "tags": [
          "AI"
        ],
        "operationId": "getAiRecoverySummary",
        "summary": "Get the latest AI recovery summary",
        "responses": {
          "200": {
            "description": "AI-oriented recovery summary.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiRecoveryResponse"
                }
              }
            }
          },
          "404": {
            "description": "No AI recovery summary is available yet.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/ai/anomalies": {
      "get": {
        "tags": [
          "AI"
        ],
        "operationId": "getAiAnomalySummary",
        "summary": "Get the latest AI anomaly summary",
        "responses": {
          "200": {
            "description": "AI-oriented anomaly summary.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiAnomalySummaryResponse"
                }
              }
            }
          },
          "404": {
            "description": "No AI anomaly summary is available yet.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/ai/context": {
      "get": {
        "tags": [
          "AI"
        ],
        "operationId": "getAiContext",
        "summary": "Get AI context for a trend window",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "7d",
                "30d"
              ]
            },
            "required": true,
            "name": "window",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "AI-oriented context summary.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiContextResponse"
                }
              }
            }
          },
          "400": {
            "description": "The window query parameter is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No AI context summary is available yet.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}